requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
SQLAlchemy==2.0.23
aiosqlite==0.20.0
matplotlib==3.9.2
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4

